+++ src/aks-preview/HISTORY.rst
@@ -12,6 +12,9 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 * Vendor new SDK and bump API version to 2025-08-02-preview.
+* * Add add-on support for `ApplicationLoadBalancer` (Application Gateway for Containers)
+  - `az aks application-load-balancer enable`: Enable Application Load Balancer add-on for an existing cluster.
+  - `az aks application-load-balancer disable`: Disable Application Load Balancer add-on for an existing cluster.
 
 19.0.0b9
 +++++++